Move MSFL::Sinatra::Helpers back into msfl/sinatra.rb to reduce complexity
